ref #5655, #5783

- `llama_batch.logits` now also indicates if embeddings are output for that token
- fix `llama_get_embeddings_ith()` to return token embeddings
- add `llama_get_embeddings_seq()` to return sequence embeddings
- fix `embedding` example to work both with BERT and non-BERT models
- fix `server` to get the resulting embeddings correctly
- add `examples/server-embd.py` helper script
- `server` supports `--pooling`

TODO:

- rename `llama_batch.logits` to `llama_batch.output` (future PR)
- [x] BERT graph does not need KV cache - remove it
- `server` should not queue partial prompts when computing embeddings (future PR)