@cebtenzzre 2024-02-29
While we're modifying embeddings.cpp - shouldn't this:
```cpp
struct llama_batch batch = llama_batch_init(n_batch, 0, n_prompts);
```
be this instead:
```cpp
struct llama_batch batch = llama_batch_init(n_batch, 0, 1);
```
Because only one sequence ID is assigned per token?