@tybalex 2024-02-29
should the embeddings from `server.cpp` get normalized too? [embedding.cpp](https://github.com/ggerganov/llama.cpp/blob/008f3fc7912bd8ecd01603fcbc580096123186cb/examples/embedding/embedding.cpp#L55) does that.