@iamlemec 2024-03-01
Regarding removing the KV cache, I think this will give a big speedup. I did a flamegraph on CUDA, and it was spending fully 50% of the time in calls to `llama_kv_cell.has_seq_id` for constructing the attention mask in `llama_set_inputs`. If you know everything is within batch though, you can construct the attention matrix from just the `batch` object.