@ggerganov 2024-03-04
> With these changes I'm getting an MSE of 1.3e-3 relative to Sentence Transformers on WikiText with nomic-embed-text-v1.f16.gguf instead of the previous 5.62e-10. Not sure what the problem is.

Does it use mean pooling? I think I got it wrong again - checking