@ggerganov 2024-03-04
So I've updated the API to support both sequence and token embeddings:

https://github.com/ggerganov/llama.cpp/blob/79e4eede2374870e915af4fe03a291a702879e52/llama.h#L657-L671

The sequence embeddings (i.e. `llama_embeddings_seq`) work only when `pooling_type != LLAMA_POOLING_TYPE_NONE`, while token embeddings are populated when `pooling_type == LLAMA_POOLING_TYPE_NONE`. In the future, we can think about adding pooling support for all models (not just BERT) - not sure if it would be useful, but it fits into the implementation logic

With this change, to make the `embedding` example work with both token and sequence emebeddings (i.e. with and without pooling), we do the following:

https://github.com/ggerganov/llama.cpp/blob/fc9af156ffa5b795bafc871554fee3ccfb800c34/examples/embedding/embedding.cpp#L52-L65

First try to get sequence embeddings. This would return NULL if pooling is disabled and we fallback to token embeddings

Hope this finally works. Let me know if you give it a try