@iamlemec 2024-03-04
Working great here! Numbers look very, very similar to earlier.

I'm a little surprised to see a slight performance degredation (around 20-30% on both CPU and CUDA). We're definitely getting gains from the kv-cache attention matrix stuff when that's a factor. But I wonder if it relates to the fact that we're breaking up the final copy into sequences? Either that or one of the smaller changes in `built_bert`. Looking into it now.