@ggerganov 2024-03-05
> I'm a little surprised to see a slight performance degredation (around 20-30% on both CPU and CUDA). We're definitely getting gains from the kv-cache attention matrix stuff when that's a factor. But I wonder if it relates to the fact that we're breaking up the final copy into sequences? Either that or one of the smaller changes in built_bert. Looking into it now.

Let me know what you find. I guess the embeddings extraction is now indeed a bit slower due to the `std::map` and loop over the batch. The actual computation seems to be the same speed as before (at least with Metal that is). You can bench it with `llama-bench` since it does not involve any embeddings extraction.

The main struggle with the API was to find a way to differentiate between token embeddings and sequence embeddings. Without the `std::map` there was an ambiguity in which embeddings you were extracting. There might be better way to do it though

The big benefit of not using the KV cache is that now we don't need to allocate a large KV cache memory buffer. For example, you can compute 32k token embeddings with just a large enough batch size (e.g. `-c 512 -b 32768`)