@iamlemec 2024-03-05
Oh nice, I didn't know about `llama-bench`. So for CPU I'm seeing ~20% performance drop, not too bad. But for CUDA I'm getting an almost 60% drop:
```
| model          | backend    | ngl | test       |             (pre) t/s |            (post) t/s |
| -------------- | ---------- | --: | ---------- | --------------------: | ---------------------:|
| bert 109M F16  | CUDA       |  99 | pp 512     | 177281.98 ± 116298.41 |  74489.67 ±  19758.33 |
| bert 109M F16  | CUDA       |   0 | pp 512     |   6824.48 ±   1151.24 |   5556.47 ±    536.14 |
```
It seems like the kv cache code relies more on `ggml_view_3d` instead of contiguous permutes, but I haven't yet found anything that yields the old performance.